changing

+++ src/main/java/beans/Node.java
@@ -1,18 +1,18 @@
 package beans;
 
 public class Node {
-    private int id;
+    private Long id;
     private String name;
     private String address;
     private double coordinateX;
     private double coordinateY;
     private String code;
 
-    public Integer getId(){
+    public Long getId(){
         return id;
     }
 
-    public void setId(Integer id){
+    public void setId(Long id){
         this.id = id;
     }
